[FIX] html_editor: don't apply batched changes to a destroyed component
ISSUE:
Prior to this commit, `embeddedState` changes could be applied on a component
after it was already destroyed.

FIX:
`embeddedState` changes are batched to allow multiple changes to be handled at
once. If the component is destroyed before the batch can be applied, it should
not be applied at all, to avoid accidentally modifying the host attribute when
it is not currently in the DOM, since it could come back with an history undo.

task-3672747

Part-of: odoo/odoo#173997
Related: odoo/enterprise#67083
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>
Co-authored-by: Adrien Schoffeniels <adsc@odoo.com>
Co-authored-by: Julien Banken <jbn@odoo.com>
Co-authored-by: Thomas Josse <thjo@odoo.com>